api/requirements.txt
fastapi==0.85.1
numpy==1.22.4
Pillow==9.2.0
tensorflow==2.10.0
uvicorn==0.14.0
gunicorn
python-multipart
